Create activity_add_route_screen.xml
The functionality that takes place in the AddRouteScreen for the portrait landscape will need to be adapted to this layout as well. Why do we have two separate layout files for this Activity? Since the portrait layout has a larger height and we wanted to fill the screen on the portrait layout for this Activity, we need a different layout to accommodate the less height we have in the landscape layout, which is the layout people who have actual keyboards on their Android phones would be using.