Create activity_add_route_screen.xml
We need two layout files for this Activity to accommodate the less height the landscape layout has compared to the portrait layout since we wanted to fill the screen in the portrait layout for this Activity. The landscape layout would be used by people who have actual keyboards on their Androids, such as the Droid 4.