Update AddRouteScreen.java
We have two Spinners: one for choosing kilometers and miles, and one for selecting the alertDistance (a conversion is made so that the alertDistance is always in miles. Right now, we need to fix the code so that all the fields are retained in startMapActivity and wiped in saveRoute (having difficulty for some reason).