Update AddRouteScreen.java
When the Save button is pressed, a notification is sent to the user that states the text message to the phone number(s) has been sent. When the user clicks on this notification, it takes them to the Main Screen and then disappears. I know this isn't what we want now, but the code is there so that when the controller is functional, it'll send a notification to the user stating the text has been sent when the distance between the GPS coordinates and the current coordinates is less than the threshold distance.