Update SettingsScreen.java
Now when hitting the Save button, a notification appears to the user stating their battery level is below the chosen threshold level and says that the location has been sent to all phone numbers they have put into the screen. While this is not our plan, this will go into the battery level controller.