Update AddRouteScreen.java
We now have an integer called timesClicked that knows how many times the Contacts button has been clicked. When that integer is odd, the first EditText is updated with the contact phone number. When the integer is even, the second EditText is updated with the contact phone number. It doesn't work yet, but when we get the methods to work, it should work as well, in theory.