Map Markers 100% Functional! :D
You can add a marker to the map by tapping, press and hold to move the
marker around. Tapping the marker will set the destination and bring
you back to the previous screen. A mini tutorial dialog was added to
the map screen also. Oh and cleaned up a little bit of old code here
and there.